requirements.txt
opencv-python
ollama
piper-tts
sounddevice
